Redefine 'abnormal aging' to remove dependency on GO:0007568.
Since 'aging' (GO:0007568) is now obsolete in GO, and does not have a
proposed replacement AFAIK, we redefine our own 'abnormal aging' so that
it no longer depends on the obsoleted GO term. Likewise for the
subclasses 'increased speed of aging' and 'decreased speed of aging'.

There is a proposal to create an 'aging rate' quality in PATO
(https://github.com/obophenotype/bio-attribute-ontology/issues/239) that
we could use to maintain a logical definition of 'abnormal aging', but
until that quality is available and with the FBcv release planned for
next week (and with it the refresh of the GO import, which will bring in
the obsoletion of GO:0007568), for now we remove the logical definition
entirely.

closes #168