Remove custom mirroring rule for RO.
The latest RO release has fixed the problem of conflicting definitions
between RO and BFO, so we no longer need to hack our RO mirror to work
around them.

We still need to work around the conflicting labels on RO coming from
the FBdv shorthands, though, because the recent fix for those shorthands
in FBdv was incomplete. We should have removed the labels on the
shorthands entirely.