Remove custom 'prepare_release' rule.
There is in fact no need in DPO to override the 'prepare_release' rule.
In other FlyBase ontologies, this was done to ensure that the "flybase
versions" of the ontologies (e.g. fly_development.obo, fly_anatomy.obo)
were properly treated as release artefacts, but in the case of DPO there
is no such "flybase version".